+++ src/XRoadLib/Attributes/IXRoadProtocolAppliable.cs
@@ -1,0 +1,9 @@
+﻿namespace XRoadLib.Attributes
+{
+    public interface IXRoadProtocolAppliable
+    {
+        bool HasAppliesToValue { get; }
+
+        XRoadProtocol AppliesTo { get; set; }
+    }
+}