+++ src/XRoadLib/Attributes/XRoadLayoutAttribute.cs
@@ -3,9 +3,11 @@
 namespace XRoadLib.Attributes
 {
     [AttributeUsage(AttributeTargets.Class)]
-    public class XRoadLayoutAttribute : Attribute
+    public class XRoadLayoutAttribute : Attribute, IXRoadProtocolAppliable
     {
-        internal XRoadProtocol? appliesTo;
+        private XRoadProtocol? appliesTo;
+
+        public bool HasAppliesToValue => appliesTo.HasValue;
 
         public XRoadProtocol AppliesTo { get { return appliesTo.GetValueOrDefault(); } set { appliesTo = value; } }